fix button things, move button observer to button file
